split ExternalArtifact into user class and config for step
