+++ src/zenml/config/step_configurations.py
@@ -26,7 +26,9 @@ from typing import (
 
 from pydantic import root_validator, validator
 
-from zenml.artifacts.external_artifact import ExternalArtifact
+from zenml.artifacts.external_artifact_config import (
+    ExternalArtifactConfiguration,
+)
 from zenml.config.base_settings import BaseSettings, SettingsOrDict
 from zenml.config.constants import DOCKER_SETTINGS_KEY, RESOURCE_SETTINGS_KEY
 from zenml.config.source import Source, convert_source_validator
@@ -166,7 +168,7 @@ class PartialStepConfiguration(StepConfigurationUpdate):
 
     name: str
     caching_parameters: Mapping[str, Any] = {}
-    external_input_artifacts: Mapping[str, ExternalArtifact] = {}
+    external_input_artifacts: Mapping[str, ExternalArtifactConfiguration] = {}
     outputs: Mapping[str, PartialArtifactConfiguration] = {}
 
     # Override the deprecation validator as we do not want to deprecate the
